refactor(api-server): extract scheduler retention/prune jobs to submodule
Move the two daily append-only retention prunes (support_tooling_events
#2531, layout_change_events #2563) — their tuning constants, the two
Scheduler methods, the pure daily-cadence gate helpers, and the gate unit
tests — out of the 2.8k-line scheduler.rs into a dedicated
services/scheduler/retention.rs submodule. scheduler.rs becomes
scheduler/mod.rs.

Pure move/extract, no behaviour change: the prune methods are attached to
Scheduler via an impl block in the child module and stay wired into
run_scheduled_tasks; the only mechanical change is bumping the two method
visibilities from private to pub(super) so the parent tick loop and the
DB integration test can still call them across the new module boundary.

Reduces churn concentration on the scheduler hot file.
